Refactored resolvers to be more modular. Added RateLimiterConfig object to graphql schema
