+++ package.json
@@ -7,8 +7,9 @@
         "test": "echo \"Error: no test specified\" && exit 1",
         "start": "cross-env NODE_ENV=development nodemon --exec ts-node --files ./server/index.ts",
         "build": "cross-env NODE_ENV=production webpack",
-        "dev": "concurrently \"cross-env NODE_ENV=development webpack serve --open\" \"cross-env NODE_ENV=development nodemon --exec ts-node --files ./server/index.ts\" ",
+        "dev": "concurrently \"cross-env NODE_ENV=development webpack serve --open\" \"cross-env NODE_ENV=production nodemon --exec ts-node --files ./server/index.ts\" ",
         "dev-client": "cross-env NODE_ENV=development webpack serve --open",
+        "dev-server": "cross-env NODE_ENV=development nodemon --exec ts-node --files ./server/index.ts",
         "lint": "eslint client server",
         "lint:fix": "eslint --fix client server",
         "prettier": "prettier --write .",
@@ -42,6 +43,7 @@
         "graphql": "^16.5.0",
         "http": "^0.0.1-security",
         "jsonwebtoken": "^8.5.1",
+        "lodash": "^4.17.21",
         "mongoose": "^6.3.3",
         "ncp": "^2.0.0",
         "randomstring": "^1.2.2",
@@ -62,6 +64,7 @@
         "@types/express": "^4.17.13",
         "@types/graphql": "^14.5.0",
         "@types/jsonwebtoken": "^8.5.8",
+        "@types/lodash": "^4.14.182",
         "@types/node": "^17.0.33",
         "@types/randomstring": "^1.1.8",
         "@types/react": "^18.0.9",
